requirements.txt
fastapi
uvicorn
python-docx
openpyxl
pandas
PyMuPDF
borb
streamlit
gradio
